+++ services/device-registry-base/src/main/java/org/eclipse/hono/deviceregistry/metrics/DeviceRegistryMetrics.java
@@ -1,0 +1,38 @@
+/**
+ * Copyright (c) 2022 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+package org.eclipse.hono.deviceregistry.metrics;
+
+import org.eclipse.hono.service.metric.Metrics;
+import org.eclipse.hono.service.metric.NoopBasedMetrics;
+
+/**
+ * Metrics for the Device Registry service.
+ */
+public interface DeviceRegistryMetrics extends Metrics {
+
+
+    /**
+     * A no-op implementation for this specific metrics type.
+     */
+    final class Noop extends NoopBasedMetrics implements DeviceRegistryMetrics {
+
+        private Noop() {
+        }
+    }
+
+    /**
+     * The no-op implementation.
+     */
+    DeviceRegistryMetrics NOOP = new Noop();
+}